requirements.txt
streamlit
python-docx
scikit-learn
spacy
